Organize ESP code for further work
