+++ wifi/main/comms.h
@@ -1,0 +1,3 @@
+#define BUF_SIZE 1024 // change this as appropriate
+
+void UART_init(void);
