+++ wifi/main/network.h
@@ -1,0 +1,4 @@
+extern const char *TAG;
+
+void wifi_init_softap(void);
+
